发言存入数据库时过滤标签,上传名词扫盲

+++ 2.2/chatDeleteAction.php
@@ -8,14 +8,27 @@ if (!isset($_SESSION['username'])){
 
 //删除该记录
 if(isset($_GET['delId'])&&isset($_GET['page'])){
+    $username=$_SESSION['username'];
+    echo $username;
     $delId=$_GET['delId'];
     $page=$_GET['page'];
     include_once 'conn.php';
-    $sql="delete from content where id=$delId";
-    $del=mysql_query($sql,$conn);
-    if ($del){
-        header("Location:chat_room.php?page=$page");
-        exit();
+    $sql="select user_id from content where id='".$delId."'";
+    $res=mysql_query($sql,$conn);
+    if ($row=mysql_fetch_assoc($res)){
+        $sql="select username from user where id='".$row['user_id']."'";
+        $res=mysql_query($sql,$conn);
+        $row=mysql_fetch_assoc($res);
+        if ($row){
+            if($row['username']==$username){
+                $sql="delete from content where id=$delId";
+                $del=mysql_query($sql,$conn);
+                if ($del){
+                    header("Location:chat_room.php?page=$page");
+                    exit();
+                }
+            }
+        }
     }
 }
 header("Location:chat_room.php?page=$page");
